Make user configurable
The script currently assumes that a user called vagrant will
be used for this environment.

This should be made configurable so that these scripts can be
run on aribitrary AMIs without modification

This patch updates the user so that it is determined by
setting the environment variable VAGRANT_USER before
running vagrant. It also updates the directory where
everything is stored to be: /home/$VAGRANT_USER (unless
the user is root in which case it will be: /root/)