Allow cassandra to start for integration tests
The current systemd config for cassandra resulted
in an issue where cassandra could never load.

It has two post actions that add a 10 second sleep, and then
another that checks to load the correct key spaces
into cassandra.

In my test cases, cassandra took more than 10 seconds to
initialize. The check in this case, failed, and killed the
running cassandra instance. This resulted in an inconsistent
data directory which prevented subsequent attempts by cassandra
to start.

This patch removes the sleep 10, and instead, runs a container
that just performs a wait-loop until cassandra is running.