Allow cassandra to start for integration tests
The current systemd config for cassandra resulted
in an issue where cassandra could never load.

It has two post actions that add a 10 second sleep, and then
another that checks to see if cassandra is loaded.

In my test cases, cassandra took more than 10 seconds to
initialize. The check in this case, failed, and killed the
running cassandra instance. This resulted in an inconsistent
data directory which prevented subsequent attempts by cassandra
to start.