Add hooks for caching some dependencies on an image
Installing everything from scratch from a bare-boned
image will add significant wall-clock time to the test runs.

This can be mitigated by allowing some things to be pre-installed
and cached on to an image which can be reused for other test runs.

This commit helps facilitate that process by allowing users (or
build systems to supply the environement variable:
    BUILD_CACHED_IMAGE

When set, this effects the normal test runs in the following ways:
  * does not copy cwd to VM
  * only runs first "install_packages" script
  * allows it to use the base image to build from