SCE-429: Initial values of cpuset.cpus are unknown.
According to cgroup documentation  we cannot assume there is any 'default'
value in cpuset.cpus cpuset.mems.
If clone_children flag in parent group is set then values of the parent
group are inherited and if the flag is not set then some properties are
empty. Therefore relying on "" or 0 as default values is not right.

Additionally cpuset.cpu_exclusive is not always settable. It can be set
only if parent group has this property set and non of the sibling group
have overlapping cpuset. Do not try to set unless you know you can.


Signed-off-by: Maciej Patelczyk <maciej.patelczyk@intel.com>