SCE-440: Snap session integration test is flaky.
Snap session is ussing net.Dial() for make sure that snapd is up and
running. Method net.Dial() is giving information that some server is
running on specified address instead of snapd responses on API calls.

To make sure that snapd is running we can use API call. If it returns
valid value, we are sure that snapd is ready.