Allow to run Caffe workload without isolations.
Be default factory, will use otherIsolation for every unknown
aggressor. Now for Caffe it will add LLC isolation, but this behavior
can be changed trough flag.
Rationale: for "real" workloads as Caffe, we want to relay on isolations
provided by Operating System and/or executor (eg. when run on kubernetes
BE is run as BestEffort class) so factory will not apply any isolations
by itself.