SCE-803: Caffe aggressor without isolation (#383) (and disable parallel tests)
* Allow to run Caffe workload without isolations.

Be default factory, will use other Isolation for every unknown
aggressor. Now for Caffe it will add LLC isolation, but this behavior
can be changed trough flag.
Rationale: for "real" workloads as Caffe, we want to relay on isolations
provided by Operating System and/or executor (eg. when run on kubernetes
BE is run as BestEffort class) so factory will not apply any isolations
by itself.


* Disabled parallel integration tests because of flake memcached (port confilit) and it saves only 1.5 minutes from total 13-14 minutes.