Moving images to clean up docs folder
Documentation is easiest to navigate on GitHub when it's only docs.
I've moved images into a new location of `/images/`. In addition, all
links use the absolute path of `/images/*.png` instead of relative
linking because files do tend to shuffle around as you get used to user
feedback on user experience.